Patch usddiff for USDZ containing crate layers
The underlying layers need to be converted to usda before they are diffed.

Previously a crate layer within a USDZ would produce the following failure when run through usddiff:

  Differences in a.usdz[basic.usdc] and b.usdz[basic.usdc]:
  'utf-8' codec can't decode byte 0xfe in position 16: invalid start byte

The change in test output is due to the ascii layer running through usdcat which it did not previously. The particulars of the diff remain the same.